+++ CHANGELOG.md
@@ -3,12 +3,13 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.7.3] - 2021-09-04
 ### Added
 - JEI support is back!
 ### Fixed
 - Crash on launch in `NerfedGear` [#427]
 - Mattocks being effective on stone [#420]
+- Armor and elytra not recognizing valid materials correctly [#418]
 - Some recipes not respecting the `allowLegacyMaterialMixing` config and allowing mixing anyway
 
 ## [2.7.2] - 2021-08-22